Broker MQTT dan REST API Sensor

+++ database/migrations/2024_12_22_094317_create_sensor_data_table.php
@@ -6,11 +6,6 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateSensorDataTable extends Migration
 {
-    /**
-     * Jalankan migrasi.
-     *
-     * @return void
-     */
     public function up()
     {
         Schema::create('sensor_data', function (Blueprint $table) {
@@ -19,17 +14,12 @@ class CreateSensorDataTable extends Migration
             $table->float('humidity');
             $table->float('light');
             $table->float('soil_moisture');
-            $table->float('tds');
-            $table->float('ph');
+            $table->integer('tds');
+            $table->integer('ph');
             $table->timestamps();
         });
     }
 
-    /**
-     * Rollback migrasi.
-     *
-     * @return void
-     */
     public function down()
     {
         Schema::dropIfExists('sensor_data');
